move api function back to utils
