+++ apps/web/src/utils/usernames.ts
@@ -1,7 +1,16 @@
+import {
+  Address,
+  Chain,
+  encodePacked,
+  keccak256,
+  namehash,
+  sha256,
+  ContractFunctionParameters,
+} from 'viem';
+import { normalize } from 'viem/ens';
 import RegistrarControllerABI from 'apps/web/src/abis/RegistrarControllerABI';
 import EARegistrarControllerAbi from 'apps/web/src/abis/EARegistrarControllerAbi';
-import { Address, Chain, encodePacked, keccak256, namehash, sha256 } from 'viem';
-import { normalize } from 'viem/ens';
+import L2ResolverAbi from 'apps/web/src/abis/L2Resolver';
 import profilePictures1 from 'apps/web/src/components/ConnectWalletButton/profilesPictures/1.svg';
 import profilePictures2 from 'apps/web/src/components/ConnectWalletButton/profilesPictures/2.svg';
 import profilePictures3 from 'apps/web/src/components/ConnectWalletButton/profilesPictures/3.svg';
@@ -26,6 +35,8 @@ import {
   IsValidIpfsUrl,
   IsValidVercelBlobUrl,
 } from 'apps/web/src/utils/urls';
+import { getBasenamePublicClient } from 'apps/web/src/hooks/useBasenameChain';
+import { USERNAME_L2_RESOLVER_ADDRESSES } from 'apps/web/src/addresses/usernames';
 
 export const USERNAME_MIN_CHARACTER_LENGTH = 3;
 export const USERNAME_MAX_CHARACTER_LENGTH = 20;
@@ -437,6 +448,68 @@ export function validateBasenameAvatarUrl(source: string): ValidationResult {
     };
   }
 }
+
+/* 
+  Fetch / Api functions
+*/
+
+// Get username `addr`
+export async function getBasenameAddress(username: BaseName) {
+  const chain = getChainForBasename(username);
+
+  try {
+    const client = getBasenamePublicClient(chain.id);
+    const ensAddress = await client.getEnsAddress({
+      name: normalize(username),
+      universalResolverAddress: USERNAME_L2_RESOLVER_ADDRESSES[chain.id],
+    });
+    return ensAddress;
+  } catch (error) {}
+}
+
+// Build a TextRecord contract request
+export function buildBasenameTextRecordContract(
+  username: BaseName,
+  key: UsernameTextRecordKeys,
+): ContractFunctionParameters {
+  const chain = getChainForBasename(username);
+  return {
+    abi: L2ResolverAbi,
+    address: USERNAME_L2_RESOLVER_ADDRESSES[chain.id],
+    args: [namehash(username), key],
+    functionName: 'text',
+  };
+}
+
+// Get a single TextRecord
+export async function getBasenameTextRecord(username: BaseName, key: UsernameTextRecordKeys) {
+  const chain = getChainForBasename(username);
+  try {
+    const client = getBasenamePublicClient(chain.id);
+    const contractParameters = buildBasenameTextRecordContract(username, key);
+    const textRecord = await client.readContract(contractParameters);
+    return textRecord as string;
+  } catch (error) {}
+}
+
+// Get a all TextRecords
+export async function getBasenameTextRecords(username: BaseName) {
+  const chain = getChainForBasename(username);
+  try {
+    const readContracts: ContractFunctionParameters[] = textRecordsKeysEnabled.map((key) => {
+      return buildBasenameTextRecordContract(username, key);
+    });
+
+    const client = getBasenamePublicClient(chain.id);
+    const textRecords = await client.multicall({ contracts: readContracts });
+
+    return textRecords;
+  } catch (error) {}
+}
+
+/* 
+  Feature flags
+*/
 
 // Force EA/GA based on env
 export const IS_EARLY_ACCESS = process.env.NEXT_PUBLIC_USERNAMES_EARLY_ACCESS == 'true';
